requirements.txt
absl-py==2.1.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.5
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dassl==0.6.3
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.16.0
flake8==3.7.9
fonttools==4.49.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.0
h5py==3.10.0
huggingface-hub==0.29.3
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.1.2
isort==4.3.21
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
littleutils==0.2.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.1
numpy==1.22.3
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
ogb==1.3.6
outdated==0.2.2
packaging==23.2
pandas==2.0.3
Pillow==9.0.1
pip==23.3.1
platformdirs==4.3.6
protobuf==4.25.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyHessian==0.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.6
tabulate==0.9.0
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==1.0.9
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.1.0
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.0.4
wesanderson==0.0.3
wheel==0.41.2
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.20.2
